+++ couriercxx/util/System.h
@@ -74,13 +74,11 @@ public:
 	static void usleep(uint64_t microseconds);
 
 	/**
-	 * Make directory
+	 * Self update
 	 *
-	 * \param[in] string path - path
-	 *
-	 * \return int - 0 if success, -1 if error
+	 * \param[in] path - path to DEB packet
 	 */
-	static int mkdir(std::string path);
+	static void selfUpdate(std::string path);
 
 	/**
 	 * Daemonize process